added double precision and fixed quotient datatype

+++ Tutorial 3/Exercise_4/main.cpp
@@ -12,7 +12,7 @@ int main()
     int sum = number1 + number2;
     int product = number1 * number2;
     int difference = number1 - number2;
-    int quotient = number1 / number2;
+    double quotient = (double) number1 / (double) number2;
     int remainder = number1 % number2;
 
     cout <<"Sum of number 1 & number 2 is: "<< sum <<endl;
